requirements.txt
google-genai==1.73.1
google-adk==1.30.0
litellm==1.83.8
openai==2.24.0
python-dotenv==1.0.1
langchain-openai==1.1.10
click==8.1.8
pydantic==2.12.5
typer==0.23.1
nemoguardrails==0.21.0
